Separate database utility functions
