+++ backend/src/api/member.rs
@@ -1,28 +1,8 @@
-use crate::models::Member;
-use diesel::{
-    query_dsl::methods::SelectDsl,
-    sql_types::{Integer, Text, Varchar},
-    Connection, PgConnection, QueryableByName, RunQueryDsl, SelectableHelper,
+use crate::{
+    database::{bind_column_value, get_connection},
+    models::Member,
 };
-use dotenv::dotenv;
-
-fn get_connection() -> Option<PgConnection> {
-    dotenv().ok();
-
-    let url = std::env::var("DATABASE_URL");
-
-    if url.is_err() {
-        return None; // FIXME Improve error message
-    }
-
-    let connection = PgConnection::establish(&url.unwrap());
-
-    if connection.is_err() {
-        return None; // FIXME Improve error message
-    }
-
-    Some(connection.unwrap())
-}
+use diesel::{query_dsl::methods::SelectDsl, sql_types::Integer, RunQueryDsl, SelectableHelper};
 
 pub fn get_all_members() -> Option<Vec<Member>> {
     use crate::schema::member::dsl::*;
@@ -56,35 +36,6 @@ pub struct ChangeRecord {
     pub value: Option<String>,
 }
 
-#[derive(QueryableByName, Debug)]
-pub struct ColumnTypeInfo {
-    #[sql_type = "Text"]
-    pub column_name: String,
-    #[sql_type = "Text"]
-    pub data_type: String,
-}
-
-fn determine_column_type(column_name: &str) -> Option<ColumnTypeInfo> {
-    let connection = get_connection();
-
-    if connection.is_none() {
-        return None;
-    }
-
-    let mut connection = connection.unwrap();
-
-    let derived_column_types = diesel::sql_query(
-        "SELECT column_name, data_type FROM information_schema.columns WHERE table_name = 'member' AND column_name = $1",
-    )
-        .bind::<Varchar, _>(column_name)
-        .load::<ColumnTypeInfo>(&mut connection);
-    if derived_column_types.is_err() || (derived_column_types.as_ref().unwrap().len() != 1) {
-        return None;
-    }
-
-    Some(derived_column_types.unwrap().pop().unwrap())
-}
-
 pub fn change_member(changes: Vec<ChangeRecord>) -> Vec<usize> {
     let connection = get_connection();
 
@@ -101,16 +52,6 @@ pub fn change_member(changes: Vec<ChangeRecord>) -> Vec<usize> {
     let mut succeeded_update_indices: Vec<usize> = Vec::new();
 
     for (index, change) in changes.iter().enumerate() {
-        let column_type = determine_column_type(change.column.as_str());
-
-        if column_type.is_none() {
-            // FIXME Either throw exception or log warning etc.
-            println!("Could not determine column type");
-            continue;
-        }
-
-        let column_type = column_type.unwrap();
-
         // FIXME Determine primary key automatically
         // FIXME Prefer query builder over raw SQL
         let unbound_update_statement = diesel::sql_query(format!(
@@ -118,9 +59,7 @@ pub fn change_member(changes: Vec<ChangeRecord>) -> Vec<usize> {
             change.column
         ));
 
-        let changed_value = change.value.as_ref();
-
-        if changed_value.is_none() {
+        if change.value.is_none() {
             // FIXME Verify whether column is nullable
             // FIXME Either throw exception or log warning etc.
             // FIXME Implement nullable case
@@ -132,20 +71,16 @@ pub fn change_member(changes: Vec<ChangeRecord>) -> Vec<usize> {
             continue;
         }
 
+        let changed_value = change.value.as_ref();
+
         let boxed_unbound_update_statement = unbound_update_statement.into_boxed();
-        let changed_value_update_statement = match column_type.data_type.as_str() {
-            "text" => boxed_unbound_update_statement.bind::<Text, _>(changed_value.unwrap()),
-            "character varying" => {
-                boxed_unbound_update_statement.bind::<Varchar, _>(changed_value.unwrap())
-            }
-            "integer" => boxed_unbound_update_statement
-                .bind::<Integer, _>(changed_value.unwrap().parse::<i32>().unwrap()),
-            _ => {
-                println!("Unknown type {}", column_type.data_type.as_str());
-                // FIXME Handle error
-                continue;
-            }
-        };
+        let changed_value_update_statement = bind_column_value(
+            change.column.as_str(),
+            changed_value.unwrap().as_str(),
+            boxed_unbound_update_statement,
+        )
+        // FIXME Improve logging and error handling
+        .expect("Could not bind column value");
         let update_statement =
             changed_value_update_statement.bind::<Integer, _>(change.membershipid);
         let update_result = update_statement.execute(&mut connection);
