+++ backend/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod api;
+
+mod database;
 mod frb_generated;
 mod models;
 mod schema;
